+++ src/main/java/company/Main.java
@@ -8,13 +8,13 @@ public class Main {
 
     public static void main(String[] args) {
 
-        Employee employee = new Employee();
-        List<Employee> allEmployees = employee.getAllEmployees();
+        Company myCompany = new Company();
+        List<Employee> allEmployees = myCompany.getAllEmployees();
 
         System.out.println(allEmployees);
-        employee.printEmployeeNames(allEmployees);
-        System.out.println(employee.getEmployeesWithSalaryMoreThan(allEmployees, 2000));
-        System.out.println(employee.getHiredEmployees(allEmployees, true));
+        myCompany.printEmployeeNames(allEmployees);
+        System.out.println(myCompany.getEmployeesWithSalaryMoreThan(allEmployees, 2000));
+        System.out.println(myCompany.getHiredEmployees(allEmployees));
